Fix error!/success! calls from BuiltinMethods
