+++ lib/manageiq/providers/workflows/builtin_methods.rb
@@ -1,11 +1,7 @@
-require_relative "builtin_result_mixin"
-
 module ManageIQ
   module Providers
     module Workflows
       class BuiltinMethods < BasicObject
-        extend BuiltinResultMixin
-
         def self.email(params = {}, _secrets = {}, _context = {})
           options = params.slice("To", "From", "Subject", "Cc", "Bcc", "Body", "Attachment").transform_keys { |k| k.downcase.to_sym }
 
@@ -22,11 +18,11 @@ module ManageIQ
           object_type, object_id = context.execution.values_at("_object_type", "_object_id")
 
           # ensure we are in a provisioning request
-          return error!({}, :cause => "Calling provision_execute on non-provisioning request: [#{object_type}]") unless object_type == "ServiceTemplateProvisionTask"
-          return error!({}, :cause => "Missing MiqRequestTask id") if object_id.nil?
+          return BuiltinRunnner.error!({}, :cause => "Calling provision_execute on non-provisioning request: [#{object_type}]") unless object_type == "ServiceTemplateProvisionTask"
+          return BuiltinRunnner.error!({}, :cause => "Missing MiqRequestTask id") if object_id.nil?
 
           miq_request_task = ::MiqRequestTask.find_by(:id => object_id.to_i)
-          return error!({}, :cause => "Unable to find MiqReqeustTask id: [#{object_id}]") if miq_request_task.nil?
+          return BuiltinRunnner.error!({}, :cause => "Unable to find MiqReqeustTask id: [#{object_id}]") if miq_request_task.nil?
 
           miq_request_task.execute
 
@@ -41,7 +37,7 @@ module ManageIQ
 
         private_class_method def self.miq_task_status!(runner_context)
           miq_task = ::MiqTask.find_by(:id => runner_context["miq_task_id"])
-          return error!(runner_context, :cause => "Unable to find MiqTask id: [#{runner_context["miq_task_id"]}]") if miq_task.nil?
+          return BuiltinRunnner.error!(runner_context, :cause => "Unable to find MiqTask id: [#{runner_context["miq_task_id"]}]") if miq_task.nil?
 
           runner_context["running"] = miq_task.state != ::MiqTask::STATE_FINISHED
 
@@ -50,7 +46,7 @@ module ManageIQ
             if runner_context["success"]
               runner_context["output"] = miq_task.message
             else
-              error!(runner_context, :cause => miq_task.message)
+              BuiltinRunnner.error!(runner_context, :cause => miq_task.message)
             end
           end
 
@@ -63,15 +59,15 @@ module ManageIQ
           case miq_request_task&.statemachine_task_status
           when nil
             reason = "Unable to find MiqRequestTask id: [#{runner_context["miq_request_task_id"]}]"
-            error!(runner_context, :cause => reason)
+            BuiltinRunnner.error!(runner_context, :cause => reason)
           when "error"
             reason = miq_request_task.message&.sub(/^Error: /, "")
-            error!(runner_context, :cause => reason)
+            BuiltinRunnner.error!(runner_context, :cause => reason)
           when "retry"
             runner_context["running"] = true
             runner_context
           when "ok"
-            success!(runner_context, :output => {"Result" => "provisioned"})
+            BuiltinRunnner.success!(runner_context, :output => {"Result" => "provisioned"})
           end
         end
       end
